requirements.txt
discord.py
pycountry
python-dotenv
gtts
wandb
openai
pynacl
